The beginnings of AI
Right now the AI can find a target, aim at the target and shoot the target. the shot applies a small amount of force

still using cubes for everything

Need to implement nav mesh so they can run around the map.

need to implement an ammo system complete with the bolt cycle times and reloading and limited ammo.

need to consolidate the scripts and change the look for component to look for layermask or tag. No need for so many scripts.